pyproject.toml
[project]
name = "patient-advocacy-agent"
version = "0.1.0"
description = "An agentic system providing dermatological triage via voice-only interface for underserved communities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.5.0",
    "transformers>=4.47.0",
    "chromadb>=0.6.0",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.6.0",
    "accelerate>=1.0.0",
    "sentencepiece>=0.2.0",
]
voice = [
    "aiortc>=1.9.0",
    "aiohttp>=3.11.0",
    "faster-whisper>=1.1.0",
    "piper-tts>=1.2.0",
]
cloud = [
    "google-genai>=1.0.0",
    "google-cloud-speech>=2.28.0",
    "google-cloud-texttospeech>=2.21.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "T20"]

[tool.ruff.lint.per-file-ignores]
"src/api/*.py" = ["B008"]  # Depends() in defaults is the FastAPI pattern
"src/auth/dependencies.py" = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-x -q --tb=short"
markers = [
    "local: tests that require MODEL_BACKEND=local and real model weights",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
